requirements.txt
numpy
Pillow
streamlit
Keras
tensorflow